requirements.txt
alembic==1.14.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
passlib==1.7.4
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
